Starting refactor, not losing progress.
