+++ atlas/config_celery.py
