+++ atlas/templates/settings.php
