--- config_flower.py
@@ -1,9 +1,0 @@
-# RabbitMQ management api
-broker_api = 'amqp://guest@localhost:5672/api/'
-
-address = 'inventory.local'
-
-# Enable debug logging
-logging = 'DEBUG'
-
-#url_prefix = 'flower'
